backend/requirements.txt
# Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
django-cors-headers>=4.0.0

# Image processing
Pillow>=9.0.0

# Environment variables
python-decouple>=3.6

# Database (for production)
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Development tools
django-debug-toolbar>=4.0.0

# Security
cryptography>=3.4.8

# Utilities
python-dateutil>=2.8.0